requirements.txt
requests~=2.28.2
colorama~=0.4.6
beautifulsoup4==4.12.2
dnspython~=2.3.0
python-whois~=0.8.0